+++ client_auth.go
@@ -21,6 +21,7 @@ import (
 	"golang.org/x/crypto/ssh/agent"
 )
 
+
 type PasswordAuthMethod struct{}
 type OidcAuthMethod struct {
 	doPKCE bool
